package*: Install 1/2 versions of same peer dep module
After running npm audit fix, npm gave 3 peer dep warnings:
ajv-keywords@2.1.1 requires a peer of ajv@^5.0.0 but none is installed
ajv-keywords@3.4.0 requires a peer of ajv@^6.9.1 but none is installed
ajv-keywords@3.4.0 requires a peer of ajv@^6.9.1 but none is installed

Since you cannot have the same module listed twice in package.json,
I decided to require ajv@^6.9.1, and press my luck with out the earlier
version. [ajv](https://ajv.js.org/) mentions moving to a new namespace
in the future, which is likely a result of this type of peer dependency
mess.